UPSTREAM: convenience helpers: vfs_get_super() and sget_fc()
the former is an analogue of mount_{single,nodev} for use in
->get_tree() instances, the latter - analogue of sget() for the
same.

These are fairly similar to the originals, but the callback signature
for sget_fc() is different from sget() ones, so getting bits and
pieces shared would be too convoluted; we might get around to that
later, but for now let's just remember to keep them in sync.  They
do live next to each other, and changes in either won't be hard
to spot.

Change-Id: I9cfbacb7151f1daf8bc575292a14ec7e2d0ce7d8
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>